chore(package): bump mapnik@4.5.5
Binaries for mapnik's dependencies couldn't be downloaded for my
Linux system on the previous mapnik version, so this commit just
raise the version to the latest, where banaries do exist. We can
drop this commit if the mapnik version needs to be the pinned to
the earlier version for other reasons.